fix: concurrency issue during update of versioned entities index
A concurrent call to get the latest version would get could get empty result during update, because the latest map is cleared before it is repopulated.

Perhaps not the best solution, but for now, synchronized access to the map during update will fix the issue.

Also, single put and remove can update the map directly.